onion: omega2p: Implement bootmenu
Implement a bootmenu protoype for the omega2p board. In normal
operation, linux will be started immediately. But holding down the reset
button on startup will launch a boot menu, which allows the user to
graphically select alternative boot options.

Note that as of now the "Web recovery" and "Flash Firmware" Boot options
are dysfunctional. Appropriate boot scripts are not yet implemented.

Signed-off-by: Philip Oberfichtner <pro@denx.de>